Kalman Filter with motion gating logic

+++ src/pnp_solver.cpp
@@ -44,7 +44,7 @@ namespace vo
                                             100, 8.0, 0.99, inlier_mask);
 
         // Make sure there are enough inliers, the pose is ok
-        if (!pose_flag || inlier_mask.rows < (size_t)(min_pnp_pts/2)) return false;
+        if (!pose_flag || inlier_mask.rows < (min_pnp_pts/2)) return false;
         // Convert the rotation vector to rotation matrix
         cv::Rodrigues(rvec, R);
         return pose_flag;
